Organizes Components into seperate dir
