+++ src/Components/EventList/Attendees.js
@@ -19,6 +19,7 @@ export default function Attendees({ event, updateEventAttendance }) {
         <div className="attendees">
           {attendees.map((attendee, index) => (
             <Attendee
+              key={attendee.id}
               attendee={attendee}
               event={event}
               updateEventAttendance={updateEventAttendance}
