+++ src/Components/EventList/Event.js
@@ -3,7 +3,7 @@ import Attendees from "./Attendees";
 export default function Event({ event, updateEventAttendance }) {
   return (
     <>
-      <li key={event.id} className="event">
+      <li key={event.id} className="events">
         <img src={event.eventImage} alt={event.name} />
         <h5>
           {event.name} {event.eventType}
